refactor(intake): split package resolution seams
Why:
- package resolution logic and tests were still clustered into oversized files
- parity work needs narrow seams and direct rule coverage rather than aggregate assertions

What:
- split intake package rules into focused models, marker, scope, duplicate, merge, review, and resolution modules
- replace oversized package tests with smaller scenario-focused files and add direct helper coverage
- keep the public intake package entry point stable through the facade module

Checks:
- uv run python -m tools.run_quality_gates --full-tests
